requirements.txt
torch
torchvision
sentence-transformers
tqdm
PyYAML
grad-cam
opencv-python-headless
matplotlib
Pillow
numpy
pyyaml
